Fix: StackOverflow with continue from try..except
Re: https://github.com/IronLanguages/main/issues/1070
was: http://ironpython.codeplex.com/workitem/34186

I had submitted this as an IronPython bug a while back. Jeff Robbins,
the CEO of LiveData, the company I work for, who was the one who
originally isolated the bug, which had first appeared in a larger
program. We then worked together on categorizing it, and I submitted the
report.

I ran the validation suite (make test) on the proposed patch, and it
actually seemed to fix a couple of errors there, at least I got errors
when I ran against the original source and they went away when I ran
again with my changes.

The rest of this is a description of the code change in the proposed
patch, its assumptions, etc.

The bug happens when a Python "continue" leaves a try..except handler
without reaching the LeaveExceptionHandlerInstruction (or maybe some
other instruction in some cases), so this misses the required return
from HandleException().

The theory of the patch is that while executing a handler, the
InstructionIndex doesn't leave a known range of instruction indices (of
course it is OK if this happens while executing a nested frame), so it
is possible to tell when such a jump out of the handler has occurred by
watching for the InstructionIndex to leave the range that comprises the
handler. This test replaces the test for the previously executed
instruction being a LeaveExceptionHandlerInstruction.

So I marked the ExceptionHandler by adding a HandlerEndIndex property,
which holds the InstructionIndex where the
LeaveExceptionHandlerInstruction is written, and changed the calls to
its constructor to provide the extra information. So I know that we have
jumped out of the handler when the index moves outside the range
handler.HandlerStartIndex..handler.HandlerEndIndex. I also know the
index of the proper LeaveExceptionHandlerInstruction, so I can execute
that instruction within HandleException() when it detects that it is
appropriate.

If that happens, then in HandleException I invoke
instructions[handler.HandlerEndIndex].Run(frame) to do anything that
needs to be done to leave the handler, before returning
ExceptionHandlingResult.Continue. I leave the InstructionIndex where it
was, so that execution continues at the appropriate spot, in the case of
the example program this would be at the top of the outer loop. That is,
it won't be the normal branch destination after
instructions[handlerEndIndex].Run(frame). My assumption is that we only
want to take the branch implied by the LeaveExceptionHandlerInstruction
(which is a kind of IndexedBranchInstruction), if we leave the handler
by reaching the LeaveExceptionHandlerInstruction in the usual way.

In that case, if the code actually makes it to handler.HandlerEndIndex,
then I just go ahead and return ExceptionHandlingResult.Continue and let
the caller actually leave the exception and take the branch. That just
made the code look a little simpler. If you would prefer to always run
the LeaveExceptionHandlerInstruction in HandleException, then I can do
that, I think it depends on where an exception raised by
LeaveExceptionHandlerInstruction.Run should be handled, and it appeared
to me that the effect would be the same. Logically, this handler is
complete at some point during the execution of
LeaveExceptionHandlerInstruction, and the structure of the code really
shouldn't depend on precisely where that point is (at least that was my
higher-level justification).

I checked for the case when there is also a "finally" clause, because I
thought that this might be a circumstance when we would indeed want to
take the branch implied by the LeaveExceptionHandlerInstruction, but it
looks like the branch for the "continue" already did the right thing,
and arranged for the branch to be postponed until the "finally" clause
is complete. I also checked a variety of nested exceptions. I didn't fix
anything with respect to "continue" out of a "finally" clause, but I
really am not sure what that should do anyway, especially if the handler
has directed control to go elsewhere (e.g. a "break" or "return" in the
handler). In any case, it is still disallowed (as in CPython).

I think these are the main weaknesses, and assumptions in the fix.

1) It requires that the code to leave the handler, the
LeaveExceptionHandlerInstruction say (or whatever it is), just be a
single instruction, but I think that restriction was there in the
original code as well. I did relax the restriction that it needs to be a
specific kind of instruction (since I replaced that test with the range
test on InstructionIndex). This "single instruction" restriction could
be relaxed in a variety of ways, if necessary.

2) In its current form, it assumes that the "continue" (or other branch)
that caused control to leave the handler did not already execute the
LeaveExceptionHandlerInstruction. If this assumption is not true, the
code here should be changed by removing the lines that invoke
instructions[handlerEndIndex].Run(frame). However, it did not appear to
me that the LeaveExceptionHandlerInstruction was being Run by the
"continue" out of the handler in any other way (and in my tests it was
only happening once per loop).

3) It also depends on the fact that while executing a frame, the
InstructionIndex doesn't leave the range of the handler while the
handler is being executed (of course it is OK if this happens while
executing within a nested frame). If this is not the case, then I think
a fix would require categorizing branches as to whether they caused an
exit from one or more handlers, so that the depth within handlers could
be maintained as an execution state of the frame. But I didn't see
anything like that (e.g. no branches out of the handler range to re-use
similar code fragments for space optimization).

4) It assumes that a "finally" clause can be executed by the caller
(normally Run(), but possibly HandleException() for nested try..except),
after the LeaveExceptionHandlerInstruction. I believe that it normally
works that way, with a branch to the "finally" clause after the
LeaveExceptionHandlerInstruction.

The main strength, I think, is that it is only a trivial change to the
original design (at least I intended it to be), the stack generally
looks the same as it looked before, except that it gets cleaned up when
the InstructionIndex leaves the region of the exception handler. It also
allows the ExceptionHandler to be cleaned up with any instruction,
looking at the code it appeared possible that it could have been cleaned
up with a LeaveFaultInstruction, but I didn't see that actually happen
and the original code didn't appear to allow for it (since it only
tested for LeaveExceptionHandlerInstruction, and not for
LeaveFaultInstruction). Perhaps it wasn't necessary to clean up the
stack in that way for LeaveFaultInstruction.

I also changed the summary comment at the top of Interpreter that
referred to the "RunInstructions" method, from context it appeared that
this method had been renamed to "Run".